magasin-ia-vision/requirements.txt
# Core & UI
PyQt6
# pyqt6-tools removed to avoid conflicts with Python 3.13
qasync>=0.27  # Asyncio loop integration for Qt

# AI & Computer Vision
ultralytics>=8.1.0  # YOLOv11 implicitly handled here
opencv-python-headless>=4.9.0
numpy>=1.26.0
torch>=2.2.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.17.0 --index-url https://download.pytorch.org/whl/cpu
supervision>=0.18.0  # ByteTrack & tools

# Data & Persistence
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
pydantic>=2.6.0

# Image & Video Processing
Pillow>=10.2.0

# Utilities
loguru>=0.7.2  # Better logging
python-dotenv>=1.0.0
pyyaml>=6.0.1
watchdog>=4.0.0

# Dev & Testing
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
